refactor: extract rank cell logic into shared utility
